+++ ft_printf/includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: wonjchoi <wonjchoi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/06 20:18:39 by wonjchoi          #+#    #+#             */
-/*   Updated: 2022/02/06 20:21:35 by wonjchoi         ###   ########.fr       */
+/*   Updated: 2022/02/07 13:00:47 by wonjchoi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int	print_persent(void);
 int	print_str(char *str);
 int	print_char(int n);
 int	print_nub(int n);
-int	print_hex(long long num, char fmt);
+int	print_hex(unsigned int num, char fmt);
 int	print_pointer(unsigned long long n);
 
 #endif
